@@ -8,11 +8,11 @@
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/pro-light-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.18",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/lab": "^4.0.0-alpha.61",
     "@material-ui/pickers": "^3.3.10",
     "@reduxjs/toolkit": "^1.8.0",
-    "axios": "^0.26.1",
+    "axios": "^0.27.2",
     "chalk": "^5.0.1",
     "clsx": "^1.1.1",
     "compressorjs": "^1.0.7",
@@ -20,8 +20,8 @@
     "lodash.omit": "^4.5.0",
     "moment": "^2.29.1",
     "print-js": "^1.6.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.7.1",
     "react-image-crop": "^8.6.12",
     "react-promise-tracker": "^2.1.0",
@@ -29,7 +29,7 @@
     "react-router-dom": "^6.2.2",
     "react-scripts": "^5.0.0",
     "redux-logger": "^3.0.6",
-    "typescript": "^4.6.2",
+    "typescript": "^4.7.4",
     "web-vitals": "^1.1.1",
     "yup": "^0.32.11"
   },
@@ -73,33 +73,33 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/humps": "^2.0.1",
     "@types/jest": "^27.4.1",
-    "@types/lodash.omit": "^4.5.6",
-    "@types/node": "^17.0.21",
+    "@types/lodash.omit": "^4.5.7",
+    "@types/node": "^17.0.29",
     "@types/react": "^18.0.14",
     "@types/react-dom": "^18.0.5",
     "@types/react-image-crop": "^8.1.3",
     "@types/react-redux": "^7.1.16",
     "@types/react-router-dom": "^5.1.7",
-    "@types/redux-logger": "^3.0.8",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
+    "@types/redux-logger": "^3.0.9",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
     "@typescript-eslint/parser": "^5.14.0",
-    "babel-plugin-import": "^1.13.3",
+    "babel-plugin-import": "^1.13.5",
     "babel-plugin-module-resolver": "^4.1.0",
     "customize-cra": "^1.0.0",
-    "eslint": "^8.2.0",
+    "eslint": "^8.18.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^16.1.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jest": "^26.1.1",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-jsx-a11y": "^6.6.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "http-proxy-middleware": "^2.0.6",
     "husky": "^7.0.0",
     "lint-staged": "^12.3.5",
-    "prettier": "^2.5.1",
+    "prettier": "^2.7.1",
     "react-app-rewired": "^2.2.1"
   }
 }
