@@ -1,18 +1,17 @@
+import App from "./App";
 import React from "react";
-import ReactDOM from "react-dom";
+import store from "@stores/store";
+import reportWebVitals from "./reportWebVitals";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import { Provider } from "react-redux";
-import store from "@stores/store";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById("root")
+const container = document.getElementById("root");
+const root = createRoot(container as Element);
+root.render(
+  <Provider store={store}>
+    <App />
+  </Provider>
 );
 
 // If you want to start measuring performance in your app, pass a function
